requirements.txt
# AI Goat - FastAPI Backend

# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy[asyncio]>=2.0.25
aiosqlite>=0.19.0
greenlet>=3.0.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0,<4.1.0

# HTTP Client
httpx>=0.26.0

# Configuration
pyyaml>=6.0.1
pydantic>=2.5.0

# AI / Embeddings
sentence-transformers>=2.3.0
chromadb>=0.4.22

# NeMo Guardrails
nemoguardrails>=0.10.0

# Development & Testing (install with: pip install -r requirements.txt)
pytest>=8.0.0
pytest-asyncio>=0.23.0
ruff>=0.4.0
pyflakes>=2.5.0
pre-commit>=3.7.0
